requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
ecdsa==0.19.1
Flask==3.1.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pyasn1==0.6.1
Pygments==2.19.2
python-jose==3.5.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.3
